[Add: 컴포넌트 분리 및 목데이터 사용과 css 수정]
- style
 Login.scss 태그 수정.
 Aside, MainLeft,Feed, Comment, AddComment, StoryFeed 컴포넌트 별 스타일 분리 작성.

- js
 Comment : useState 및 useEffect 사용하여 목데이터 및 댓글 추가 기능 구현.
 Main, Aside, MainLeft, Feed, Comment, AddComment, StoryFeed 컴포넌트 분리작성.
