+++ src/pages/haesoo/Login/Login.scss
@@ -118,7 +118,7 @@
         }
 
         .find-password-container {
-          a {
+          p {
             font-weight: 500;
             color: #385085;
           }
@@ -141,7 +141,7 @@
           font-size: 14px;
         }
 
-        a {
+        span {
           margin: 0px 2px;
           color: #0095f6;
           font-size: 14px;
