+++ src/pages/haesoo/Main/Aside/Aside.scss
@@ -1,0 +1,159 @@
+.aside-haesoo {
+  .user-container {
+    display: flex;
+    flex-direction: column;
+    width: 310px;
+    padding: 0px 10px;
+
+    .user-wrapper {
+      display: flex;
+      align-items: center;
+      width: 100%;
+      padding: 20px 10px;
+
+      .user-wrapper__image {
+        width: 50px;
+        border-radius: 50%;
+        cursor: pointer;
+
+        img {
+          width: 100%;
+          height: 50px;
+          border-radius: 50%;
+        }
+      }
+
+      .user-wrapper__information {
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        width: 140px;
+        margin-left: 20px;
+        padding: 15px 4px;
+
+        .user-wrapper__id {
+          width: 100%;
+          font-weight: 600;
+          text-align: left;
+          cursor: pointer;
+        }
+
+        .user-wrapper__name {
+          margin-top: 4px;
+          color: #9c9c9c;
+        }
+      }
+
+      .user-wrapper__mode {
+        padding-left: 30px;
+        color: #0095f6;
+        font-weight: 600;
+        font-size: 12px;
+        cursor: pointer;
+      }
+    }
+
+    .recommend-wrapper {
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      width: 100%;
+
+      .recommend-wrapper__text {
+        display: flex;
+        justify-content: space-between;
+        width: 100%;
+        padding: 0px 20px;
+        margin-bottom: 20px;
+
+        p {
+          color: #9c9c9c;
+          font-weight: 600;
+          font-size: 14px;
+        }
+        button {
+          color: #231f20;
+          font-weight: 600;
+          font-size: 13px;
+          cursor: pointer;
+        }
+      }
+
+      .recommend-wrapper__contents {
+        display: flex;
+        flex-direction: column;
+        width: 100%;
+        padding: 0px 20px;
+
+        .recommend-wrapper__follow {
+          display: flex;
+          justify-content: space-between;
+          width: 100%;
+          height: 40px;
+          margin-bottom: 20px;
+
+          .follow__image {
+            width: 30px;
+            height: 30px;
+            border-radius: 50%;
+            cursor: pointer;
+
+            img {
+              width: 100%;
+              height: 100%;
+              border-radius: 50%;
+            }
+          }
+
+          .follow__user {
+            display: flex;
+            flex-direction: column;
+            width: 180px;
+            cursor: pointer;
+
+            .user-id {
+              margin-bottom: 8px;
+              font-weight: 600;
+            }
+
+            .user-follow {
+              color: #9c9c9c;
+              font-size: 12px;
+            }
+          }
+
+          .follow__button {
+            color: #0095f6;
+            font-weight: 600;
+            font-size: 12px;
+            cursor: pointer;
+          }
+        }
+      }
+    }
+
+    .main-footer {
+      display: flex;
+      flex-direction: column;
+      width: 100%;
+
+      .footer-tags {
+        width: 100%;
+        padding: 20px 20px;
+        margin-bottom: 10px;
+        color: #c0c0c0;
+        font-size: 12px;
+
+        span {
+          margin-right: 2px;
+        }
+      }
+
+      .footer-copyright {
+        padding: 0px 20px;
+        color: #c0c0c0;
+        font-size: 12px;
+      }
+    }
+  }
+}
