+++ src/pages/haesoo/Main/MainLeft/Feed/Comment/AddComment/AddComment.scss
@@ -1,0 +1,35 @@
+.feed-comment__add {
+  display: flex;
+  flex-direction: column;
+
+  .comment-wrapper {
+    position: relative;
+    display: flex;
+    width: 100%;
+    padding: 0px 20px;
+    margin-bottom: 10px;
+  }
+
+  .comment-id {
+    width: auto;
+    margin-right: 10px;
+    font-weight: 600;
+    font-size: 16px;
+    cursor: pointer;
+  }
+
+  .comment-text {
+    width: auto;
+  }
+
+  .delete-button {
+    position: absolute;
+    right: 25px;
+    cursor: pointer;
+    opacity: 0.3;
+
+    &:hover {
+      opacity: 1;
+    }
+  }
+}
