+++ src/pages/haesoo/Main/MainLeft/Feed/Feed.scss
@@ -1,0 +1,169 @@
+.feed-container {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  background-color: #fff;
+  border: 1px solid #dbdbdb;
+  border-radius: 4px;
+
+  .feed-header {
+    position: relative;
+    display: flex;
+    align-items: center;
+    width: 100%;
+    padding: 20px 20px;
+
+    .header__image {
+      width: 28px;
+      height: 28px;
+      margin-right: 20px;
+      border: 1px solid #dbdbdb;
+      border-radius: 50%;
+    }
+
+    .header__id {
+      width: auto;
+      font-size: 18px;
+      font-weight: 600;
+    }
+
+    .header__button {
+      all: unset;
+      position: absolute;
+      right: 20px;
+      color: #231f20;
+      font-size: 8px;
+      cursor: pointer;
+    }
+  }
+
+  .feed-image {
+    width: 100%;
+    height: 780px;
+
+    .image__upload {
+      width: 100%;
+      height: 760px;
+    }
+
+    .image__upload {
+      img {
+        width: 100%;
+        height: 100%;
+      }
+    }
+
+    .image__buttons {
+      all: unset;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      width: 100%;
+      height: 20px;
+
+      p {
+        width: 7px;
+        height: 7px;
+        margin: 2px;
+        background-color: #9c9c9c;
+        border-radius: 50%;
+      }
+    }
+  }
+
+  .feed-icons {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    width: 100%;
+    padding: 10px 0px;
+
+    .feed-icons__left {
+      width: 50%;
+      left: 0px;
+      padding: 0px 20px;
+      cursor: pointer;
+
+      img {
+        width: 20px;
+        margin-right: 10px;
+      }
+    }
+
+    .feed-icons__right {
+      width: 50%;
+      right: 0;
+      padding: 0px 10px;
+      text-align: right;
+
+      img {
+        width: 20px;
+        margin-right: 10px;
+      }
+    }
+  }
+  .feed-information {
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+    padding: 10px 20px;
+
+    .feed-likes {
+      width: 100%;
+      font-size: 14px;
+      font-weight: 600;
+      margin-bottom: 20px;
+    }
+
+    .feed-user {
+      display: flex;
+      align-items: center;
+
+      .feed-user__id {
+        width: auto;
+        font-weight: 600;
+        font-size: 16px;
+        cursor: pointer;
+      }
+
+      .feed-user__text {
+        width: auto;
+        margin-left: 10px;
+      }
+    }
+  }
+
+  .feed-comment {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    width: 100%;
+    padding: 20px 20px;
+    border-top: 1px solid #dbdbdb;
+
+    .feed-comment__left {
+      position: relative;
+      width: 100%;
+
+      i {
+        font-size: 18px;
+      }
+
+      input {
+        all: unset;
+        width: 80%;
+        margin-left: 10px;
+        font-size: 14px;
+      }
+
+      .feed-comment__button {
+        position: absolute;
+        right: 5px;
+        color: #0095f6;
+        font-weight: 600;
+        font-size: 14px;
+        cursor: pointer;
+      }
+    }
+  }
+}
