+++ src/pages/haesoo/Main/MainLeft/MainLeft.scss
@@ -1,0 +1,11 @@
+.main-left {
+  display: flex;
+  width: 935px;
+  margin: 0px 100px;
+  padding-top: 84px;
+
+  .contents-container {
+    width: 600px;
+    padding: 0px 10px 20px 20px;
+  }
+}
