+++ src/pages/haesoo/Main/MainLeft/StoryFeed/StoryFeed.js
@@ -1,0 +1,51 @@
+import React from 'react';
+import './StoryFeed.scss';
+
+export default function StoryFeed() {
+  return (
+    <div className="story-container-haesoo">
+      <div className="story-container__contents">
+        <div className="story__image">
+          <img alt="story profile" src="/images/haesoo/story_images/1.JPG" />
+        </div>
+        <div className="story__id">onnuk_</div>
+      </div>
+      <div className="story-container__contents">
+        <div className="story__image">
+          <img alt="story profile" src="/images/haesoo/story_images/2.JPG" />
+        </div>
+        <div className="story__id">singmulstore</div>
+      </div>
+      <div className="story-container__contents">
+        <div className="story__image">
+          <img alt="story profile" src="/images/haesoo/story_images/3.JPG" />
+        </div>
+        <div className="story__id">dlwldms</div>
+      </div>
+      <div className="story-container__contents">
+        <div className="story__image">
+          <img alt="story profile" src="/images/haesoo/story_images/7.JPG" />
+        </div>
+        <div className="story__id">_jeongjaehyun</div>
+      </div>
+      <div className="story-container__contents">
+        <div className="story__image">
+          <img alt="story profile" src="/images/haesoo/story_images/5.JPG" />
+        </div>
+        <div className="story__id">cy_1.018</div>
+      </div>
+      <div className="story-container__contents">
+        <div className="story__image">
+          <img alt="story profile" src="/images/haesoo/story_images/4.JPG" />
+        </div>
+        <div className="story__id">bona_is_cute</div>
+      </div>
+      <div className="story-container__contents">
+        <div className="story__image">
+          <img alt="story profile" src="/images/haesoo/story_images/6.JPG" />
+        </div>
+        <div className="story__id">ninicho</div>
+      </div>
+    </div>
+  );
+}
