+++ src/pages/haesoo/Main/MainLeft/StoryFeed/StoryFeed.scss
@@ -1,0 +1,36 @@
+.story-container-haesoo {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  background-color: #fff;
+  padding: 20px;
+  margin-bottom: 24px;
+  border: 1px solid #dbdbdb;
+  border-radius: 4px;
+
+  .story-container__contents {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    width: 70px;
+
+    .story__image {
+      width: 50px;
+      margin-bottom: 4px;
+
+      img {
+        width: 100%;
+        height: 50px;
+        border-radius: 50%;
+      }
+    }
+  }
+  .story__id {
+    width: 100%;
+    padding: 3px 0px;
+    font-size: 13px;
+    text-align: center;
+  }
+}
